# docs/swagger.yaml
basePath: /
definitions:
  models.Application:
    properties:
      ai_score:
        type: integer
      applied_at:
        type: string
      candidate_id:
        type: string
      candidate_name:
        description: Для отображения в списке рекрутера
        type: string
      id:
        type: string
      status:
        type: string
      vacancy_id:
        type: string
    type: object
...
host: https://ai-recruiting-4l5g.onrender.com
info:
  contact: {}
  description: API сервис для автоматизации рекрутинга.
  title: AI Recruiting API
  version: "1.0"
paths:
...
  /applications/{id}:
    get:
      description: 'Получает список всех откликов, привязанных к конкретному ID вакансии.
        (Маршрут: /:id/vacancy, где id - это ID вакансии)'
      parameters:
      - description: ID вакансии
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Список заявок
          schema:
            items:
              $ref: '#/definitions/models.Application'
            type: array
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Список заявок на вакансию
      tags:
      - applications
  /applications/{id}/ai-data:
    get:
      description: Получает вердикт, скоринг и распознанный текст резюме
      parameters:
      - description: ID заявки (application_id)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Данные анализа
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Данные не найдены
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Результаты анализа ИИ
      tags:
      - applications
  /applications/{id}/status:
    patch:
      consumes:
      - application/json
      description: 'Обновляет статус конкретного отклика (например: ''Интервью'',
        ''Оффер'', ''Отказ'') по его ID.'
      parameters:
      - description: ID заявки (application_id)
        in: path
        name: id
        required: true
        type: string
      - description: Новый статус
        in: body
        name: input
        required: true
        schema:
          properties:
            status:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Status updated
          schema:
            type: string
        "400":
          description: Ошибка валидации
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Ошибка БД
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Изменение статуса заявки
      tags:
      - applications
  /applications/{id}/vacancy:
    get:
      description: 'Возвращает подробную информацию о конкретном отклике по его ID.
        (Маршрут: /:id, где id - это ID заявки)'
      parameters:
      - description: ID заявки (application_id)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Объект заявки
          schema:
            $ref: '#/definitions/models.Application'
        "404":
          description: Заявка не найдена
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Полная информация о заявке
      tags:
      - applications
...
  /templates/{id}/generate:
    post:
      consumes:
      - application/json
      description: Генерирует текст сообщения и ссылку на чат по шаблону
      parameters:
      - description: ID шаблона
        in: path
        name: id
        required: true
        type: string
      - description: Данные кандидата
        in: body
        name: input
        required: true
        schema:
          properties:
            candidate_name:
              type: string
            telegram_username:
              type: string
            vacancy_title:
              type: string
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Текст и ссылка
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Ошибка входных данных
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Шаблон не найден
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Генерация сообщения для Telegram
      tags:
      - recruiter_tools
...
swagger: "2.0"